// src/primer-api/model/nodeFlavorBoxBody.ts
/**
 * Generated by orval v6.15.0 🍺
 * Do not edit manually.
 * Primer backend API
 * A backend service implementing a pedagogic functional programming language.
 * OpenAPI spec version: 0.7
 */

export type NodeFlavorBoxBody = typeof NodeFlavorBoxBody[keyof typeof NodeFlavorBoxBody];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NodeFlavorBoxBody = {
  Pattern: 'Pattern',
} as const;