// src/primer-api/model/nodeFlavorPrimBody.ts
/**
 * Generated by orval v6.19.1 🍺
 * Do not edit manually.
 * Primer backend API
 * A backend service implementing a pedagogic functional programming language.
 * OpenAPI spec version: 0.7
 */

export type NodeFlavorPrimBody = typeof NodeFlavorPrimBody[keyof typeof NodeFlavorPrimBody];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NodeFlavorPrimBody = {
  PrimCon: 'PrimCon',
  PrimPattern: 'PrimPattern',
} as const;